TS: reducers/status/browser
